Fix LGTM alerts and ensure that the module_redis_connection_error_message_printf_(non)critical functions are checked with the printf arguments validation

+++ src/module/redis/module_redis.c
@@ -245,7 +245,7 @@ bool module_redis_process_data(
                         module_redis_connection_error_message_printf_noncritical(
                                 connection_context,
                                 "ERR unknown command `%.*s` with `%d` args",
-                                command_length,
+                                (int)command_length,
                                 command_data,
                                 connection_context->reader_context.arguments.count - 1);
                         continue;
